# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/convert/eur-to-msats": {
            "get": {
                "description": "Converts a Euro amount to millisatoshis using current exchange rate",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "conversion"
                ],
                "summary": "Convert EUR to millisatoshis",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Amount in EUR",
                        "name": "amount",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API Key for authentication",
                        "name": "api_key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.ConversionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Verifies connectivity to a specified wallet or all wallets if none specified",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Check health of wallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet ID to check. If not provided, checks all wallets.",
                        "name": "wallet_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.HealthResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/nwc_payment": {
            "post": {
                "description": "Transfer funds from one wallet to another using EUR amount",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Make an NWC payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key for authentication",
                        "name": "api_key",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Payment Information",
                        "name": "payment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.NwcPaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.NwcPaymentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.ConversionResponse": {
            "type": "object",
            "properties": {
                "euro_amount": {
                    "type": "number"
                },
                "msat_amount": {
                    "type": "integer"
                }
            }
        },
        "main.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "main.HealthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "wallets": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "boolean"
                    }
                }
            }
        },
        "main.NwcPaymentRequest": {
            "type": "object",
            "required": [
                "euro_amount",
                "recipient",
                "sender"
            ],
            "properties": {
                "euro_amount": {
                    "type": "number",
                    "example": 0.000001
                },
                "recipient": {
                    "type": "string",
                    "example": "WALLET_VRATA_KRKE"
                },
                "sender": {
                    "type": "string",
                    "example": "WALLET_JOSIP"
                }
            }
        },
        "main.NwcPaymentResponse": {
            "type": "object",
            "properties": {
                "amount_msats": {
                    "type": "integer"
                },
                "euro_amount": {
                    "type": "number"
                },
                "fees_paid": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "recipient_balance": {
                    "type": "integer"
                },
                "sender_balance": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}